Added connection functionality.
The reason the client wasn't connecting is because it was trying to connect before everything in the page was parsed -- so I added a DOMContentLoaded event to make sure that doesn't happen anymore. Now it's possible just to include a javascript file in a <script> tag rather than having to embed it physically on the page.